Create `CacheProcessor.cc` source file
This is more of the work to split up files containing multiple classes.
